pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "django-llm"
version = "1.0.0"
description = "A beautiful unified LLM interface for Django with persistence and streaming"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    {name = "Chris Hasinski", email = "krzysztof.hasinski@gmail.com"}
]
keywords = ["django", "llm", "openai", "anthropic", "gemini", "chatgpt", "claude", "ai", "azure", "ollama"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 5.1",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "django>=4.2",
    "httpx>=0.25.0",
]

[project.urls]
Homepage = "https://github.com/khasinski/django-llm"
Documentation = "https://github.com/khasinski/django-llm#readme"
Repository = "https://github.com/khasinski/django-llm"
Changelog = "https://github.com/khasinski/django-llm/blob/main/CHANGELOG.md"
Issues = "https://github.com/khasinski/django-llm/issues"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-django>=4.5",
    "pytest-asyncio>=0.21",
    "vcrpy>=5.0",
    "ruff>=0.1.0",
]

[tool.setuptools.packages.find]
include = ["django_llm*"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "example_project.settings"
python_files = ["test_*.py"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["django_llm"]
